Encapsulate binding+drawing into rendermesh
