+++ Include/Mesh/Vertex.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <glm/vec3.hpp>
 
 struct Vertex {
